requirements.txt
discord.py
beautifulsoup4
aiohttp
requests
praw